refactor: update to NativeScript 6.0 (#220)
* refactor: bump template versions to {N} 6.0

* refactor: set `markingMode` to `none`

* refactor: remove `moduleId` from component metadata

{N} 6.0 uses only webpack for building. When building Angular apps with
Webpack,the `module.id` is inserted automatically and we don't need to
set it manually. There's even a loader in `nativescript-dev-webpack` which removes the moduleIds from the source code.

In short, the `moduleId: module.id` is useful only for builds without webpack. NativeScript 6.0 doesn't support such builds. That's why we don't need moduleId in the templates anymore.