fix: update version of `@nativescript/schematics` to `0.8.0` inside templates
Currently `@nativescript/schematics@~0.7.0` is added as dependency to the project, so when a new project is created an old version of `@nativescript/schematics` is installed. Updating the version to `@8.0.0` will lead to an another problem that a version that is still not published to `npm` will be referenced within the project. In such a situation, the project should be created using `--skip-install` flag e.g `ng new --collection=@nativescript/schematics myApp --shared --skip-install`, after that the next version should be installed in the project before executing `tns run`.